Rework "lastTelecommandId" into "lastRequestId"
Apart from changing the name I also changed the type to
`MessageTypeIdFields`. To prevent circular dependencies I then had to
move `MessageTypeIdFields` from `RfProtocol/` to `Vocabulary/`. Of
course, I also changed the type in the `TelemetryRecord`.
