+++ Sts1CobcSw/RfProtocols/Vocabulary.cpp
@@ -4,15 +4,6 @@
 namespace sts1cobcsw
 {
 template<std::endian endianness>
-auto SerializeTo(void * destination, MessageTypeIdFields const & fields) -> void *
-{
-    destination = SerializeTo<endianness>(destination, fields.serviceTypeId);
-    destination = SerializeTo<endianness>(destination, fields.messageSubtypeId);
-    return destination;
-}
-
-
-template<std::endian endianness>
 auto SerializeTo(void * destination, Parameter const & parameter) -> void *
 {
     destination = SerializeTo<endianness>(destination, parameter.parameterId);
@@ -21,19 +12,8 @@ auto SerializeTo(void * destination, Parameter const & parameter) -> void *
 }
 
 
-template auto SerializeTo<std::endian::big>(void *, MessageTypeIdFields const &) -> void *;
-template auto SerializeTo<std::endian::little>(void *, MessageTypeIdFields const &) -> void *;
 template auto SerializeTo<std::endian::big>(void *, Parameter const &) -> void *;
 template auto SerializeTo<std::endian::little>(void *, Parameter const &) -> void *;
-
-
-template<std::endian endianness>
-auto DeserializeFrom(void const * source, MessageTypeIdFields * fields) -> void const *
-{
-    source = DeserializeFrom<endianness>(source, &fields->serviceTypeId);
-    source = DeserializeFrom<endianness>(source, &fields->messageSubtypeId);
-    return source;
-}
 
 
 template<std::endian endianness>
@@ -45,10 +25,6 @@ template<std::endian endianness>
 }
 
 
-template auto DeserializeFrom<std::endian::big>(void const *, MessageTypeIdFields *)
-    -> void const *;
-template auto DeserializeFrom<std::endian::little>(void const *, MessageTypeIdFields *)
-    -> void const *;
 template auto DeserializeFrom<std::endian::big>(void const *, Parameter *) -> void const *;
 template auto DeserializeFrom<std::endian::little>(void const *, Parameter *) -> void const *;
 }
