+++ Sts1CobcSw/RfProtocols/Vocabulary.hpp
@@ -10,22 +10,6 @@
 
 namespace sts1cobcsw
 {
-struct MessageTypeIdFields
-{
-    std::uint8_t serviceTypeId = 0;
-    std::uint8_t messageSubtypeId = 0;
-
-    friend constexpr auto operator==(MessageTypeIdFields const &, MessageTypeIdFields const &)
-        -> bool = default;
-};
-
-
-template<>
-inline constexpr std::size_t serialSize<MessageTypeIdFields> =
-    totalSerialSize<decltype(MessageTypeIdFields::serviceTypeId),
-                    decltype(MessageTypeIdFields::messageSubtypeId)>;
-
-
 struct Parameter
 {
     enum class Id : std::uint8_t
@@ -49,12 +33,7 @@ inline constexpr std::size_t serialSize<Parameter> =
 
 
 template<std::endian endianness>
-[[nodiscard]] auto SerializeTo(void * destination, MessageTypeIdFields const & fields) -> void *;
-template<std::endian endianness>
 [[nodiscard]] auto SerializeTo(void * destination, Parameter const & parameter) -> void *;
-template<std::endian endianness>
-[[nodiscard]] auto DeserializeFrom(void const * source, MessageTypeIdFields * fields)
-    -> void const *;
 template<std::endian endianness>
 [[nodiscard]] auto DeserializeFrom(void const * source, Parameter * parameter) -> void const *;
 }
