+++ Tests/ManualTests/ReportFrames.test.cpp
@@ -141,7 +141,7 @@ private:
                 .epsIsWorking = 1,
                 .flashIsWorking = 1,
                 .rfIsWorking = 1,
-                .lastTelecommandIdWasInvalid = 1,
+                .lastMessageTypeIdWasInvalid = 1,
                 .lastTelecommandArgumentsWereInvalid = 1,
                 .nTotalResets = 1U,
                 .nResetsSinceRf = 2U,
@@ -173,7 +173,7 @@ private:
                 .nGoodTransferFrames = 57U,
                 .nBadTransferFrames = 58U,
                 .lastFrameSequenceNumber = 59U,
-                .lastTelecommandId = 60U
+                .lastMessageTypeId = {60U, 61U},
             };
             auto report = sts1cobcsw::HousekeepingParameterReport(record);
             WriteToFileAsFrame(report, outputDir + reportName);
